% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_cancel_analysis.R
\name{cancel_analysis}
\alias{cancel_analysis}
\title{Cancel an analysis by specifying its name.}
\usage{
cancel_analysis(key = NULL, login_email, analysis_name)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{login_email}{character, Your StreetLight login email}

\item{analysis_name}{character, The analysis name}
}
\value{
If successful, a success message, otherwise a warning message.
}
\description{
Cancel an analysis by specifying its name.
}
