% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analysis_review.R
\name{check_analysis_review}
\alias{check_analysis_review}
\title{Check the review status of a analysis in the StreetLight API.}
\usage{
check_analysis_review(analysis_name = NULL, uuid = NULL, key = NULL)
}
\arguments{
\item{analysis_name}{character, The analysis name}

\item{uuid}{character, UUID for a given analysis.}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}
}
\value{
If successful, a list with two items
Otherwise, an httr2 response.
}
\description{
Check the review status of a analysis in the StreetLight API.
}
