% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_api_key_access}
\alias{check_api_key_access}
\title{Check API key existence}
\usage{
check_api_key_access(key = NULL)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}
}
\value{
API key
}
\description{
Check API key existence
}
\keyword{internal}
