% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_range.R
\name{check_date_range}
\alias{check_date_range}
\title{Check available date ranges for a given travel mode type}
\usage{
check_date_range(
  key = NULL,
  travel_mode_type = "All_Vehicles_CVD_Plus",
  country = "US"
)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{travel_mode_type}{character, 'All_Vehicles_AGPS', 'All_Vehicles_By_Weight', 'All_Vehicles_CVD_Plus', 'All_Vehicles_LBS_Plus', 'Bicycle', 'Bus', 'Pedestrian', 'Rail', 'Truck'.
Default is 'All Vehicles'.}

\item{country}{character, one of \code{"US"} or \code{"CA"}. Default is \code{"US"}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{travel_mode_type}, \code{start_date},
and \code{end_date}.
}
\description{
Useful for looking up most recent available data to avoid
errors in analysis creation.
}
\examples{
\dontrun{
library(streetlightR)
check_date_range()
}

}
