% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_streetlight_api.R
\name{check_streetlight_api}
\alias{check_streetlight_api}
\title{Test connectivity and authentication to the API server.}
\usage{
check_streetlight_api(key = NULL)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}
}
\value{
a PASS or FAIL message
}
\description{
Test connectivity and authentication to the API server.
}
\note{
Take care of the rate limit on the API. For quick checks, a 1 second break (\code{Sys.sleep(1)}) is usually sufficient.
}
