% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_zone_size}
\alias{check_zone_size}
\title{Check that there are under 7,000 zones in a given zone set}
\usage{
check_zone_size(zones)
}
\arguments{
\item{zones}{GeoJSON feature collection \strong{or} sf object
where the features are either lines or polygons and columns include
"name" and "geometry".}
}
\value{
If too many zones, error.
}
\description{
Check that there are under 7,000 zones in a given zone set
}
\keyword{internal}
