% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_streetlight_analysis.R
\name{create_streetlight_analysis}
\alias{create_streetlight_analysis}
\title{Create a StreetLight analysis, which analyzes traffic or activity with respect to the analysis Zones.}
\usage{
create_streetlight_analysis(
  login_email,
  key = NULL,
  analysis_type,
  analysis_name,
  travel_mode_type = "All_Vehicles_CVD_Plus",
  output_type = "index",
  travel_mode_data = "",
  description = "",
  origin_zone_set,
  destination_zone_set = NA,
  middle_zone_set = NA,
  aadt_zone_set = NA,
  calibration_zone_set = NA,
  geography_type = "",
  zone_intersection_type = "",
  date_ranges = list(start_date = "01/01/2021", end_date = "12/31/2021"),
  day_types = "All Days|17,Average Weekday|15,Average Weekend Day|67",
  day_parts =
    "All Day|0023,Early AM|0005,Peak AM|0609,Mid-Day|1014,Peak PM|1518,Late PM|1923,Recreation Hours|0819",
  vehicle_weight = "",
  segment_types = list(),
  trip_attributes = FALSE,
  trip_speed_bins = "0-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90+",
  trip_duration_bins =
    "0-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90-100,100-110,110-120,120-130,130-140,140-150,150+",
  trip_length_bins =
    "0-1,1-2,2-5,5-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90-100,100+",
  trip_circuity_bins = "1-2,2-3,3-4,4-5,5-6,6+",
  enable_speed_percentile = FALSE,
  speed_percentile_bins = "5,15,85,95",
  traveler_attributes = FALSE,
  enable_home_work_locations = FALSE,
  hwl_enable_visitor = FALSE,
  hwl_enable_resident = FALSE,
  hwl_enable_worker = FALSE,
  metric_type = "segment",
  aadt_year = "",
  aadt_calibration_year = "",
  tags = "streetlightR",
  is_ui_enabled = FALSE,
  enable_visualization = FALSE,
  enable_15min = FALSE,
  enable_upsampling = TRUE,
  is_massive_queue = FALSE,
  enable_completion_email = FALSE,
  osm_ids = NA,
  unit_of_measurement = "miles"
)
}
\arguments{
\item{login_email}{character, Your StreetLight login email}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{analysis_type}{character, What type of analysis to run.
Options are 'AADT', 'Network_OD', 'Network_Performance', 'OD_Analysis', 'OD_MF_Analysis', 'OD_Preset_Geography', 'Segment_Analysis', 'Top_Routes_OD', 'Top_Routes_ZA', 'Zone_Activity_Analysis'.}

\item{analysis_name}{character, The analysis name}

\item{travel_mode_type}{character, 'All_Vehicles_AGPS', 'All_Vehicles_By_Weight', 'All_Vehicles_CVD_Plus', 'All_Vehicles_LBS_Plus', 'Bicycle', 'Bus', 'Pedestrian', 'Rail', 'Truck'.
Default is 'All Vehicles'.}

\item{output_type}{character, One of 'index', 'trip_counts', 'volume', 'zone_counts'.
Default is 'index'.}

\item{travel_mode_data}{character, Defines the truck weight classes that analysis metrics describe.
Only available when \code{travel_mode_type} is "Truck".
One of 'LIGHT_DUTY', 'MEDIUM_DUTY_HEAVY_DUTY'. Default is "".}

\item{description}{character, Optional analysis description}

\item{origin_zone_set}{character, The name of uploaded zone set to use as the
origin in an origin-destination analysis or the main zone in a zone activity analysis}

\item{destination_zone_set}{character, The name of uploaded zone set to use as
the destination in an origin-destination analysis.}

\item{middle_zone_set}{character, The name of uploaded zone set to use as
the middle filter in an origin-destination with middle filter analysis.}

\item{aadt_zone_set}{character, The name of uploaded zone set to use in an analysis with AADT output.}

\item{calibration_zone_set}{character, name of uploaded zone set with calibration.
Required when creating an analysis with Zone Counts output.}

\item{geography_type}{character, Required in O-D to Pre-set Geography analyses.
This property is a string of one of 'blkgrp', 'da', 'taz', 'zip'.}

\item{zone_intersection_type}{character, one of \code{all_trips_for_zone} or \code{trips_by_pass_through_setting}.
Applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{date_ranges}{list, a list of date ranges. Each date range is an object
containing a pair of MM/DD/YYYY dates, with the \code{start_date} key containing the start of the date range,
and the \code{end_date} key containing the end of the date range. Both \code{start_date} and \code{end_date} are inclusive.
Default is all 2021 months.
When \code{travel_mode_type} is \code{Bicycle} or \code{Pedestrian}, each date range must consist of only full months.}

\item{day_types}{character, a comma-separated list of day types in the analysis. Each day type has a name separated by the vertical bar
from the start day of week to the end day of week (1 for Monday through 7 for Sunday).
Analysis must define All Days as \code{17} (Monday through Sunday), and they must define values for
Average Weekday and Average Weekend Day. Default includes All Days, Average Weekday, and Average Weekend Day.}

\item{day_parts}{character, a comma-separated list of day parts in the analysis. Each day part has a name separated
by the vertical bar from the start hour and end hour from 00 (midnight) to 23 (11 PM).
For example, \code{"All Day|0023"} ranges from midnight (00:00) to 11:59 PM. Analyses must define All Day as 0023 (midnight to midnight).
Default is \code{"All Day|0023,Early AM|0005,Peak AM|0609,Mid-Day|1014,Peak PM|1518,Late PM|1923,Recreation Hours|0819"}.}

\item{vehicle_weight}{character, deprecated. See \code{travel_mode_data} parameter.}

\item{segment_types}{list, the categories of road segments to be analyzed.
Only required for Top Routes between Origins and Destinations Analysis or Top Routes for Zones Analysis.
must contain at least one of 'Motorway', 'Primary', 'Residential', 'Secondary', 'Tertiary', 'Trunk'}

\item{trip_attributes}{logical, This property controls whether the analysis results will include
the add-on trip attribute Metrics: trip time distribution, trip length distribution,
trip speed distribution, and trip circuity distribution.
If you specify this, you can also customize the following properties: \code{trips_speed_bins},
\code{trip_duration_bins}, \code{trip_length_bins},
\code{trip_circuity_bins}. Default is \code{FALSE}.}

\item{trip_speed_bins}{character, the default ranges with a a comma-separated list of speed ranges.
Default is \code{"0-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90+"}}

\item{trip_duration_bins}{character, the default ranges with a comma-separated
list of trip duration, or travel time, in minutes. Default is
\code{"0-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90-100,100-110,110-120,120-130,130-140,140-150,150+"}}

\item{trip_length_bins}{character, the default ranges with a a comma-separated list
of trip length ranges in the specified \code{unit_of_measurement}.
Default value is \code{"0-1,1-2,2-5,5-10,10-20,20-30,30-40,40-50,50-60,60-70,70-80,80-90,90-100,100+"}}

\item{trip_circuity_bins}{character,  the default ranges with a a comma-separated list of trip circuity ranges.
Trip circuity is the average ratio of trip length to the direct distance between the start and endpoints of the trip.
Default value is \code{"1-2,2-3,3-4,4-5,5-6,6+"}}

\item{enable_speed_percentile}{logical, whether to divide metrics into speed percentile bins.
Default is \code{FALSE}.}

\item{speed_percentile_bins}{character, If \code{enable_speed_percentile} is TRUE,
you can specify a comma-separated list of integers defining speed
percentiles (must be multiples of 5). Percentile values must be between 0
and 100. Default percentile values are "5,15,85,95".}

\item{traveler_attributes}{logical, whether the analysis results will include the add-on traveler attribute metrics.
Traveler attribute metrics include traveler demographics and simple trip purpose.
(visitor income, education, race, and family status) are included in the Metric results. Default is \code{FALSE}.}

\item{enable_home_work_locations}{logical, whether the Analysis results will include Home and Work Locations metrics.
If this is \code{TRUE}, then one of the \code{hwl_...} parameters must be \code{TRUE}. Default is \code{FALSE}.}

\item{hwl_enable_visitor}{logical, whether the Analysis results will include visiting trips that neither reside or work in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{hwl_enable_resident}{logical, whether the Analysis results will include trips that reside in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{hwl_enable_worker}{logical, whether the Analysis results will include trips that work in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{metric_type}{character, If you are running a Network Performance analysis, select the type of metrics.
Spot metrics describe vehicles at the midpoint of a road segment;
segment metrics describe vehicles over the length of a road segment.
Must be one of 'segment', 'spot'.
Default is "segment".}

\item{aadt_year}{character, a four digit year between \code{2000} and \code{2099}.
Required integer when creating an AADT Analysis.}

\item{aadt_calibration_year}{character, a four digit year between \code{2000} and \code{2099}.
Required integer when creating an AADT Analysis.}

\item{tags}{list, tag names created within an Organization to associate with the created Analysis.
Default is \code{list("streetlightR")}}

\item{is_ui_enabled}{This allows analysis results to be downloaded and visualized
though the UI as well as the API. Should be used sparingly.}

\item{enable_visualization}{logical, whether to enable visualization on the StreetLight platform}

\item{enable_15min}{logical, whether the Analysis will analyze in 15-minute day parts.}

\item{enable_upsampling}{logical, whether the Analysis will process with upsampling if it meets the necessary thresholds.
This setting only applies to analysis with \code{enable_15min} enabled.}

\item{is_massive_queue}{logical, whether the Analysis will process alongside other high volume Analyses in order to optimize calculation time.}

\item{enable_completion_email}{logical, whether the analysis will send an email upon completion. Default is \code{FALSE}}

\item{osm_ids}{list, array of OSM IDs for network analyses.}

\item{unit_of_measurement}{character, unit of measure for trip attributes.
One of \code{"miles"} or \code{"km"}. Default is \code{"miles"}}
}
\value{
If successful, a list with the analysis name, status, and universal unique ID (uuid).
}
\description{
\code{create_streetlight_analysis()} interfaces with the 'analyses' endpoint of the StreetLight Insights API.

Not all \code{analysis_type} and \code{travel_mode_type} combinations are supported. Check the
API \href{https://developer.streetlightdata.com/docs/defining-analyses#supported-analysis-type-and-travel-mode-combinations}{documentation site} for
the most up to date configuration.
}
\examples{
\dontrun{

library(streetlightR)

create_streetlight_analysis(
  login_email = "you@email.com",
  analysis_type = "Zone_Activity_Analysis",
  analysis_name = "v231218-VMT 2019 month 3 Volume Trip Trav Attr",
  travel_mode_type = "All_Vehicles_CVD_Plus",
  output_type = "volume",
  origin_zone_set = "7-County Metro + 6 MPO Municipalities",
  trip_attributes = FALSE,
  traveler_attributes = FALSE,
  date_ranges = list(
    start_date = "01/01/2019",
    end_date = "01/31/2019"
  ),
  day_types = "All Days|17,Average Weekday|15,Average Weekend Day|67",
  day_parts = "All Day|0023, Morning|0709, Midday|1113, Evening|1618",
  tags = list("streetlightR")
)
}

}
