% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tag.R
\name{create_tag}
\alias{create_tag}
\title{Create a tag}
\usage{
create_tag(key = NULL, login_email, tag_name)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{login_email}{character, Your StreetLight login email}

\item{tag_name}{character, the tag you want to create}
}
\value{
If successful, a success message.
}
\description{
Tags are useful for organizing analyses by project, organization, user, or anything else!
}
\seealso{
Other tags: 
\code{\link{list_available_tags}()}
}
\concept{tags}
