% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_shapefile.R
\name{get_analysis_shapefile}
\alias{get_analysis_shapefile}
\title{Get Shapefiles by analysis name}
\usage{
get_analysis_shapefile(
  analysis_name = NULL,
  key = NULL,
  shapefile,
  analysis_name_ = NULL
)
}
\arguments{
\item{analysis_name}{character, analysis name}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{shapefile}{character, which Shapefile to fetch. Availability is
dependent on the analysis type and configuration.
See \code{\link[=create_streetlight_analysis]{create_streetlight_analysis()}} for more details.}

\item{analysis_name_}{Deprecated, use \code{analysis_name} parameter.}
}
\value{
a tibble with the requested metric
}
\description{
Get Shapefiles by analysis name
}
