% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_tags.R
\name{list_available_tags}
\alias{list_available_tags}
\title{List all available tags}
\usage{
list_available_tags(key = NULL)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{analysis_num}, \code{created_by}, \code{created_date},
and \code{tag_name}.
}
\description{
List all available tags
}
\seealso{
Other tags: 
\code{\link{create_tag}()}
}
\concept{tags}
