% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_analysis_id.R
\name{lookup_analysis_id}
\alias{lookup_analysis_id}
\title{Look-up analysis name or UUID}
\usage{
lookup_analysis_id(analysis_name = NULL, uuid = NULL, key = NULL)
}
\arguments{
\item{analysis_name}{character, The analysis name}

\item{uuid}{character, UUID for a given analysis.}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}
}
\value{
character, the corresponding UUID or analysis name
}
\description{
Provide either \code{analysis_name} or \code{uuid} and the function will return the
corresponding \code{analysis_name} or \code{uuid}
}
\examples{
\dontrun{

library(streetlightR)

lookup_analysis_id(analysis_name = "Pt Douglas Regional Parking Lot Summer 22")
#> "89ce3700-872e-40d5-981b-019fe0663884"

lookup_analysis_id(uuid = "89ce3700-872e-40d5-981b-019fe0663884")
#> "Pt Douglas Regional Parking Lot Summer 22"
}
}
