% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_analysis_tag.R
\name{remove_analysis_tag}
\alias{remove_analysis_tag}
\title{Remove a specific tag from an analysis}
\usage{
remove_analysis_tag(key = NULL, login_email, analysis_name, tag_name)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{login_email}{character, Your StreetLight login email}

\item{analysis_name}{character, The analysis name}

\item{tag_name}{character, the tag you want to create}
}
\value{
If successful, a success message
}
\description{
Remove a specific tag from an analysis
}
