% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_osm_ids.R
\name{search_osm_ids}
\alias{search_osm_ids}
\title{Search OSM IDs}
\usage{
search_osm_ids(
  key = NULL,
  polygon = NULL,
  tract_id = NULL,
  zip_id = NULL,
  blockgroup_id = NULL,
  point = NULL,
  radius = NULL,
  unit = "mi"
)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{polygon}{GeoJSON feature collection \strong{or} sf object of type}

\item{tract_id}{character, 11 digit census tract identifier}

\item{zip_id}{character, 5-digit ZIP code}

\item{blockgroup_id}{character, 12 digit block group identifier}

\item{radius}{}

\item{unit, }{buffer distance unit of measurement.
One of \code{"mi"} or \code{"km"}. Default is \code{"mi"}}
}
\value{
If successful, a list with the API response and data frame of intersecting OSM segment IDs.
Columns include \code{"osm_id"}, \code{"road_classification"}, and \code{"osm_segment_count"}.
}
\description{
Using a polygon, census tract, census blockgroup, zip code, or point, search StreetLight's OSM database.
}
\examples{
\dontrun{

library(streetlightR)

zip_response <- search_osm_ids(zip_id = "55104")

}
}
