% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetlight_api_key.R
\name{streetlight_api_key}
\alias{streetlight_api_key}
\title{Install a StreetLight API key in your \code{.Renviron} for repeated use}
\usage{
streetlight_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The API key provided to you from StreetLight formated in quotes.}

\item{overwrite}{logical, If this is set to TRUE, it will overwrite an existing STREETLIGHT_API_KEY
that you already have in your \code{.Renviron} file.}

\item{install}{logical, if \code{TRUE}, will install the key in your \code{.Renviron}
file for use in future sessions. Default is \code{FALSE}.}
}
\description{
Install a StreetLight API key in your \code{.Renviron} for repeated use
}
\note{
Adapted from \url{tidycensus::census_key}
}
\examples{
\dontrun{
streetlight_key("111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("STREETLIGHT_API_KEY")
}

\dontrun{
# If you need to overwrite an existing key:
streetlight_key("111111abc", overwrite = TRUE, install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("STREETLIGHT_API_KEY")
}
}
