% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetlightR-package.R
\docType{package}
\name{streetlightR-package}
\alias{streetlightR}
\alias{streetlightR-package}
\title{streetlightR: Interact with StreetLight Data's API in R}
\description{
R wrapper for StreetLight Insights API. IMPORTANT: This package is not affiliated with nor maintained by StreetLight Data \url{https://www.streetlightdata.com/}. The package is meant solely as a tool for working with StreetLight's API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://congenial-waffle-78e26bc8.pages.github.io/}
  \item Report bugs at \url{https://github.com/Metropolitan-Council/streetlightR/issues}
}

}
\author{
\strong{Maintainer}: Liz Roten \email{liz.roten@metc.state.mn.us} (\href{https://orcid.org/0000-0002-5346-3549}{ORCID})

Authors:
\itemize{
  \item Raven McKnight \email{raven.mcknight@metc.state.mn.us}
  \item Ellen Esch \email{ellen.esch@metc.state.mn.us} (\href{https://orcid.org/0000-0002-4253-0910}{ORCID})
}

Other contributors:
\itemize{
  \item Metropolitan Council [copyright holder]
}

}
\keyword{internal}
