% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_analysis.R
\name{tag_analysis}
\alias{tag_analysis}
\title{Add a specific tag to an analysis}
\usage{
tag_analysis(key = NULL, login_email, analysis_name, tag_name)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{login_email}{character, Your StreetLight login email}

\item{analysis_name}{character, The analysis name}

\item{tag_name}{character, the tag you want to create}
}
\value{
If successful, a success message
}
\description{
The tag must already exist on the server. You can create a new tag with \code{\link[=create_tag]{create_tag()}}.
}
