% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_parameters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate parameters}
\usage{
validate_parameters(param, value)
}
\arguments{
\item{param}{character, parameter as character}

\item{value}{any, parameter value}
}
\value{
Throws error if value is invalid
}
\description{
Validate parameters
}
\examples{

validate_parameters("analysis_type", "Zone_Activity_Analysis")

}
\keyword{internal}
