% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analysis_status.R
\name{check_analysis_status}
\alias{check_analysis_status}
\title{Check the status of a analysis in the StreetLight API.}
\usage{
check_analysis_status(analysis_name = NULL, key = NULL, analysis_name_ = NULL)
}
\arguments{
\item{analysis_name}{character, analysis name}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{analysis_name_}{Deprecated, use \code{analysis_name} parameter.}
}
\value{
a the parsed response with the analysis availability, metrics, name, and UUID
}
\description{
Check the status of a analysis in the StreetLight API.
}
