% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analysis_status.R
\name{check_analysis_status}
\alias{check_analysis_status}
\title{Check the status of a analysis in the StreetLight API.}
\usage{
check_analysis_status(analysis_name = NULL, key = NULL, analysis_name_ = NULL)
}
\arguments{
\item{analysis_name}{character, analysis name}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{analysis_name_}{Deprecated, use \code{analysis_name} parameter.}
}
\value{
If successful, a list with two itesms
- \code{analyses}
- \code{status}

\if{html}{\out{<div class="sourceCode">}}\preformatted{Otherwise, an httr2 response. 
}\if{html}{\out{</div>}}
}
\description{
Check the status of a analysis in the StreetLight API.
}
