% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_range.R
\name{check_date_range}
\alias{check_date_range}
\title{Check available date ranges for a given travel mode type}
\usage{
check_date_range(key = NULL, travel_mode_type = "All_Vehicles", country = "US")
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{travel_mode_type}{character, \code{All_Vehicles},  \code{Truck}, \code{Bicycle}, or \code{Pedestrian}. Default is \verb{All Vehicles}}

\item{country}{character, one of \code{"US"} or \code{"CA"}. Default is \code{"US"}.}
}
\value{
\url{tibble::tibble} with columns \code{travel_mode_type}, \code{start_date},
and \code{end_date}.
}
\description{
Useful for looking up most recent available data to avoid
errors in analysis creation.
}
\examples{
\dontrun{
library(streetlightR)
check_date_range()
}

}
