% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_streetlight_analysis.R
\name{create_streetlight_analysis}
\alias{create_streetlight_analysis}
\title{Create a StreetLight analysis, which analyzes traffic or activity with respect to the analysis Zones.}
\usage{
create_streetlight_analysis(
  login_email,
  key = NULL,
  analysis_type,
  analysis_name,
  travel_mode_type = "All_Vehicles",
  output_type = "index",
  description = "",
  origin_zone_set,
  destination_zone_set = NA,
  middle_zone_set = NA,
  aadt_zone_set = NA,
  calibration_zone_set = NA,
  geography_type = "",
  zone_intersection_type = "",
  date_ranges = list(start_date = "01/01/2019", end_date = "12/31/2019"),
  day_types = "All Days|17,Average Weekday|15,Average Weekend Day|67",
  day_parts =
    "All Day|0023,Early AM|0005,Peak AM|0609,Mid-Day|1014,Peak PM|1518,Late PM|1923,Recreation Hours|0819",
  vehicle_weight = "",
  segment_types = list(),
  trip_attributes = FALSE,
  traveler_attributes = FALSE,
  enable_home_work_locations = FALSE,
  hwl_enable_visitor = FALSE,
  hwl_enable_resident = FALSE,
  hwl_enable_worker = FALSE,
  aadt_year = "",
  aadt_calibration_year = "",
  tags = "streetlightR",
  is_ui_enabled = FALSE,
  enable_visualization = FALSE,
  enable_15min = FALSE,
  enable_upsampling = TRUE,
  is_massive_queue = FALSE
)
}
\arguments{
\item{login_email}{character, Your StreetLight login email}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{analysis_type}{character, What type of analysis to run. Options are \code{OD_Analysis}, \code{OD_MF_Analysis},
\code{Zone_Activity_Analysis},\code{OD_Preset_Geography}, \code{Segment_Analysis}, \code{AADT}, \code{Top_Routes_OD},
\code{Top_Routes_ZA}, or \code{Traffic_Diagnostics}.}

\item{analysis_name}{character, The analysis name}

\item{travel_mode_type}{character, \code{All_Vehicles},  \code{Truck}, \code{Bicycle}, or \code{Pedestrian}. Default is \verb{All Vehicles}}

\item{output_type}{character, One of \code{volume}, \code{trip_counts}, \code{aadt}, \code{index}, or \code{zone_counts}. Default is \code{index}. For Traffic Diagnostics, \code{index} is allowed.}

\item{description}{character, Optional analysis description}

\item{origin_zone_set}{character, The name of uploaded zone set to use as the origin in an origin-destination analysis or the main zone in a zone activity analysis}

\item{destination_zone_set}{character, The name of uploaded zone set to use as the destination in an origin-destination analysis.}

\item{middle_zone_set}{character, The name of uploaded zone set to use as the middle filter in an origin-destination with middle filter analysis.}

\item{aadt_zone_set}{character, The name of uploaded zone set to use in an analysis with AADT output.}

\item{calibration_zone_set}{character, name of uploaded zone set with calibration. Required when creating an Analysis with Zone Counts output}

\item{geography_type}{character, Required in O-D to Pre-set Geography analyses. This property is a string of one of \code{zip}, \code{taz}, \code{da}, or \code{blkgrp}.}

\item{zone_intersection_type}{character, one of \code{all_trips_for_zone} or \code{trips_by_pass_through_setting}.
Applies only to Zone Activity Analysis with Home and Work Locations metrics enabled) r}

\item{date_ranges}{list, a list of date ranges. Each date range is an object containing a pair of MM/DD/YYYY dates, with the \code{start_date} key containing the start of the date range,
and the \code{end_date} key containing the end of the date range. Both \code{start_date} and \code{end_date} are inclusive. Default is all 2019 months.
When \code{travel_mode_type} is \code{Bicycle} or \code{Pedestrian}, each date range must consist of only full months.}

\item{day_types}{character, a comma-separated list of day types in the analysis. Each day type has a name separated by the vertical bar
from the start day of week to the end day of week (1 for Monday through 7 for Sunday).
Analysis must define All Days as \code{17} (Monday through Sunday), and they must define values for
Average Weekday and Average Weekend Day. Default includes All Days, Average Weekday, and Average Weekend Day.}

\item{day_parts}{character,A comma-separated list of day parts in the analysis. Each day part has a name separated
by the vertical bar from the start hour and end hour from 00 (midnight) to 23 (11 PM).
For example, \verb{All Day|0023} ranges from midnight (00:00) to 11:59 PM. Analyses must define All Day as 0023 (midnight to midnight).
Default includes All Day, Recreation Hours (8am-8pm), Park Hours I, II, and III.}

\item{vehicle_weight}{character, one of \code{Medium}, \code{Heavy}, \verb{Medium,Heavy} or \code{null}.
Whether metric results for the analysis are broken down
by vehicle weight class (medium duty, heavy duty) for commercial vehicles.
If its value is null, commercial vehicle results are not broken down by vehicle weight class.
Required when creating a Top Routes between Origins and Destinations Analysis or Top Routes for Zones Analysis.}

\item{segment_types}{list, must contain at least one of \code{Motorway}, \code{Trunk}, \code{Primary}, \code{Secondary}, \code{Tertiary}, \code{Residential}.}

\item{trip_attributes}{logical, This property controls whether the analysis results will include the add-on trip attribute Metrics:
trip time distribution, trip length distribution, trip speed distribution, and trip circuity distribution.
If you specify this, you can also customize the following properties: - \code{trip_duration_bins} -  \code{trip_length_bins} -
\code{trips_speed_bins} - \code{trip_circuity_bins}. Default is FALSE.}

\item{traveler_attributes}{logical, whether the analysis results will include the add-on traveler attribute metrics.
Traveler attribute metrics include traveler demographics and simple trip purpose.
(visitor income, education, race, and family status) are included in the Metric results. Default is FALSE.}

\item{enable_home_work_locations}{logical, whether the Analysis results will include Home and Work Locations metrics.
If this is \code{TRUE}, then one of the \code{hwl_...} parameters must be \code{TRUE}. Default is \code{FALSE}.}

\item{hwl_enable_visitor}{logical, whether the Analysis results will include visiting trips that neither reside or work in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{hwl_enable_resident}{logical, whether the Analysis results will include trips that reside in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{hwl_enable_worker}{logical, whether the Analysis results will include trips that work in the zone.
It applies only to Zone Activity Analysis with Home and Work Locations metrics enabled.}

\item{aadt_year}{character, one of \code{2017}, \code{2018}, or \code{2019}. Required integer creating an AADT Analysis.}

\item{aadt_calibration_year}{character, one of \code{2017}, \code{2018}, or \code{2019}. Required integer creating an AADT Analysis.}

\item{tags}{list, tag names created within an Organization to associate with the created Analysis. Default is \code{list("streetlightR")}}

\item{is_ui_enabled}{This allows analysis results to be downloaded and visualized though the UI as well as the API. Should be used sparingly.}

\item{enable_visualization}{logical, whether to enable visualization on the StreetLight platform}

\item{enable_15min}{logical, whether the Analysis will analyze in 15-minute day parts.}

\item{enable_upsampling}{logical, whether the Analysis will process with upsampling if it meets the necessary thresholds.
This setting only applies to analysis with \code{enable_15min} enabled.}

\item{is_massive_queue}{logical, whether the Analysis will process alongside other high volume Analyses in order to optimize calculation time.}
}
\value{
if successful, a list with the analysis name, status, and universal unique ID (uuid).
}
\description{
Create a StreetLight analysis, which analyzes traffic or activity with respect to the analysis Zones.
}
