% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_data.R
\name{get_analysis_data}
\alias{get_analysis_data}
\title{Get Metric result content for an Analysis by Analysis Name}
\usage{
get_analysis_data(
  analysis_name = NULL,
  key = NULL,
  metric,
  analysis_name_ = NULL
)
}
\arguments{
\item{analysis_name}{character, analysis name}

\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{metric}{character, which metric to fetch. Metric availability is
dependent on the analysis type and configuration.
See \code{\link[=create_streetlight_analysis]{create_streetlight_analysis()}} for more details.}

\item{analysis_name_}{Deprecated, use \code{analysis_name} parameter.}
}
\value{
a tibble with the requested metric
}
\description{
Get Metric result content for an Analysis by Analysis Name
}
