% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetlight_insight.R
\name{streetlight_insight}
\alias{streetlight_insight}
\title{StreetLight Insights API wrapper}
\usage{
streetlight_insight(key, endpoint)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{endpoint}{character, API endpoint. One of
'analyses', 'analyses/download/name', 'analyses/status', 'analyses/tags', 'debug/echo', 'osm_ids/search', 'tags', 'tags/remove_analyses_tag', 'tags/tag_analyses', 'zone_sets', 'zone_sets/search'.}
}
\value{
\code{\link[httr2:request]{httr2::request()}} with path and key specifications
}
\description{
StreetLight Insights API wrapper
}
\keyword{internal}
