% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetlight_insight.R
\name{streetlight_insight}
\alias{streetlight_insight}
\title{StreetLight Insights API wrapper}
\usage{
streetlight_insight(key, endpoint)
}
\arguments{
\item{key}{character, StreetLight API key. Default is \code{NULL}.}

\item{endpoint}{character, API endpoint. One of
\code{c("analyses/status",  "zone_sets", "debug/echo", "analyses/download/name")}}
}
\value{
\url{httr2::request} with path and key specifications
}
\description{
StreetLight Insights API wrapper
}
\keyword{internal}
