% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_parameters}
\alias{valid_parameters}
\title{Valid API parameters}
\format{
named list of valid values for each parameter.
Includes analysis_type, travel_mode_type, travel_mode_data, output_type, unit_of_measurement, unit, segment_types, geography_type, zone_intersection_type, metric_type, endpoint, country.
}
\source{
\href{https://developer.streetlightdata.com/}{StreetLight Developer Portal}
}
\usage{
valid_parameters
}
\description{
Valid API parameters
}
\keyword{datasets}
